requirements.txt
aspose-cells==21.2.1
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
JPype1==1.2.1
MarkupSafe==2.0.1
openpyxl==3.0.7
Werkzeug==2.0.1
